Scraping/forex/requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.30
botocore==1.35.30
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
jmespath==1.0.1
lxml==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
